+++ Framework-Release/PanicARLib.framework/Versions/A/Headers/PARController+Config.h
@@ -22,24 +22,6 @@
 
 
 
-#pragma mark - Debugging
-
-/*! enable console (will show console display if possible)
- @remarks DEBUG only */
-- (void)enableConsole;
-/*! show console display
- @remarks DEBUG only */
-- (void)showConsole;
-/*! hide console display
- @remarks DEBUG only */
-- (void)hideConsole;
-/*! disable console
- @remarks DEBUG only */
-- (void)disableConsole;
-/*! check if console is enabled
- @return YES if console is enabled
- @remarks DEBUG only */
-- (BOOL)isConsoleEnabled;