+++ PanicARDemo/PanicARDemo/AboutViewController.m
@@ -15,7 +15,6 @@
     self = [super initWithNibName:nibNameOrNil bundle:nibBundleOrNil];
     if (self) {
         self.title = NSLocalizedString(@"About", @"About");
-        self.tabBarItem.image = [UIImage imageNamed:@"second"];
         self.navigationItem.title = self.title;
         
     }