+++ PanicARDemo/PanicARDemo/ConfigViewController.mm
@@ -17,7 +17,6 @@
     self = [super initWithNibName:nibNameOrNil bundle:nibBundleOrNil];
     if (self) {
         self.title = NSLocalizedString(@"Config", @"Config");
-        self.tabBarItem.image = [UIImage imageNamed:@"second"];
         self.navigationItem.title = self.title;
         self.navigationItem.leftBarButtonItem = [[UIBarButtonItem alloc] initWithBarButtonSystemItem:UIBarButtonSystemItemRefresh target:self action:@selector(doCompassTracking:)];
         self.navigationItem.leftBarButtonItem.enabled = NO;